% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read GTDB metadata files}
\usage{
read_metadata(file)
}
\arguments{
\item{file}{A path to a file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Read GTDB metadata files
}
