% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{separate_gtdb_taxonomy}
\alias{separate_gtdb_taxonomy}
\title{Separate the GTDB Taxonomy column into one column per rank}
\usage{
separate_gtdb_taxonomy(metadata)
}
\arguments{
\item{metadata}{GTDB metadata table.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Separate the GTDB Taxonomy column into one column per rank
}
